Fixed remaining address and undefined sanitizer fuzzer errors.
The pcidsk_fuzzer exposed a bug in VSICurlHandle::Read(). The values were: iterOffset=280, nOffsetToDownload=0, osRegion.size()=272. I didn't try to understand why the file offset seems to be past the end of file. But this change can do no harm and it removes a vunerability. I also ran the gtiff_fuzzer for a day and there were no errors.

The only pcidsk_fuzzer errors left are slow-unit errors which I'm not too sure how to fix those.